extract IntUiBridgeComboBox
