+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/IntUiBridge.kt
@@ -269,50 +269,6 @@ private fun readSimpleListItemStyle() =
             ),
     )
 
-private fun readDefaultComboBoxStyle(): ComboBoxStyle {
-    val normalBackground = retrieveColorOrUnspecified("ComboBox.background")
-    val nonEditableBackground = retrieveColorOrUnspecified("ComboBox.nonEditableBackground")
-    val normalContent = retrieveColorOrUnspecified("ComboBox.foreground")
-    val normalBorder = retrieveColorOrUnspecified("Component.borderColor")
-    val focusedBorder = retrieveColorOrUnspecified("Component.focusedBorderColor")
-
-    val colors =
-        ComboBoxColors(
-            background = normalBackground,
-            nonEditableBackground = nonEditableBackground,
-            backgroundDisabled = retrieveColorOrUnspecified("ComboBox.disabledBackground"),
-            backgroundFocused = normalBackground,
-            backgroundPressed = normalBackground,
-            backgroundHovered = normalBackground,
-            content = normalContent,
-            contentDisabled = retrieveColorOrUnspecified("ComboBox.disabledForeground"),
-            contentFocused = normalContent,
-            contentPressed = normalContent,
-            contentHovered = normalContent,
-            border = normalBorder,
-            borderDisabled = retrieveColorOrUnspecified("Component.disabledBorderColor"),
-            borderFocused = focusedBorder,
-            borderPressed = focusedBorder,
-            borderHovered = normalBorder,
-        )
-
-    val minimumSize = JBUI.CurrentTheme.ComboBox.minimumSize().toDpSize()
-    val arrowWidth = JBUI.CurrentTheme.Component.ARROW_AREA_WIDTH.dp
-    return ComboBoxStyle(
-        colors = colors,
-        metrics =
-            ComboBoxMetrics(
-                arrowAreaSize = DpSize(arrowWidth, minimumSize.height),
-                minSize = DpSize(minimumSize.width + arrowWidth, minimumSize.height),
-                cornerSize = componentArc,
-                contentPadding = retrieveInsetsAsPaddingValues("ComboBox.padding"),
-                borderWidth = DarculaUIUtil.LW.dp,
-                maxPopupHeight = Dp.Unspecified,
-            ),
-        icons = ComboBoxIcons(chevronDown = AllIconsKeys.General.ChevronDown),
-    )
-}
-
 private fun readDefaultDropdownStyle(menuStyle: MenuStyle): DropdownStyle {
     val normalBackground = retrieveColorOrUnspecified("ComboBox.background")
     val normalContent = retrieveColorOrUnspecified("ComboBox.foreground")
@@ -912,5 +868,5 @@ private fun readIconButtonStyle(): IconButtonStyle =
             ),
     )
 
-private val componentArc: CornerSize
+internal val componentArc: CornerSize
     get() = CornerSize(DarculaUIUtil.COMPONENT_ARC.dp / 2)
